Extract CheckpointAcceptnaceService and migrate CheckpointBlockValidator to F[_]
