+++ src/main/scala/org/constellation/consensus/ConsensusManager.scala
@@ -6,6 +6,7 @@ import cats.implicits._
 import com.typesafe.config.Config
 import io.chrisdavenport.log4cats.SelfAwareStructuredLogger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
+import org.constellation.checkpoint.{CheckpointAcceptanceService, CheckpointService}
 import org.constellation.consensus.Consensus._
 import org.constellation.p2p.{Cluster, DataResolver, PeerData, PeerNotification}
 import org.constellation.primitives.Schema.{CheckpointCache, Id, NodeType, SignedObservationEdgeCache}
@@ -21,6 +22,7 @@ class ConsensusManager[F[_]: Concurrent](
   transactionService: TransactionService[F],
   concurrentTipService: ConcurrentTipService[F],
   checkpointService: CheckpointService[F],
+  checkpointAcceptanceService: CheckpointAcceptanceService[F],
   soeService: SOEService[F],
   messageService: MessageService[F],
   observationService: ObservationService[F],
@@ -81,7 +83,7 @@ class ConsensusManager[F[_]: Concurrent](
           roundData._3,
           new DataResolver,
           transactionService,
-          checkpointService,
+          checkpointAcceptanceService,
           messageService,
           observationService,
           remoteSender,
@@ -125,7 +127,7 @@ class ConsensusManager[F[_]: Concurrent](
             _ =>
               stopBlockCreationRound(
                 StopBlockCreationRound(error.roundId, None, error.transactions, error.observations)
-              )
+            )
           )
           .flatMap(_ => Sync[F].raiseError[ConsensusInfo[F]](error))
       case unknown =>
@@ -199,7 +201,7 @@ class ConsensusManager[F[_]: Concurrent](
           arbitraryMsgs,
           new DataResolver,
           transactionService,
-          checkpointService,
+          checkpointAcceptanceService,
           messageService,
           observationService,
           remoteSender,
@@ -282,7 +284,7 @@ class ConsensusManager[F[_]: Concurrent](
             dao.threadSafeMessageMemPool.activeChannels
               .get(message.signedMessageData.data.channelId)
               .foreach(_.release())
-        )
+      )
     )
 
   def cleanUpLongRunningConsensus: F[Unit] =
@@ -363,7 +365,8 @@ class ConsensusManager[F[_]: Concurrent](
         ConfigUtil.config.getString("constellation.consensus.arbitrary-tx-distance-base")
       ).getOrElse("hash") match {
         case "id" =>
-          (id: Id, tx: Transaction) => Distance.calculate(tx.src.address, id)
+          (id: Id, tx: Transaction) =>
+            Distance.calculate(tx.src.address, id)
         case "hash" =>
           (id: Id, tx: Transaction) =>
             val xorIdTx = Distance.calculate(tx.hash, id)
