+++ src/main/scala/org/constellation/primitives/CheckpointBlock.scala
@@ -123,27 +123,6 @@ case class CheckpointBlock(
 
   def transactionsValid(implicit dao: DAO): Boolean = transactions.nonEmpty && transactions.forall(_.valid)
 
-  // TODO: Return checkpoint validation status for more info rather than just a boolean
-
-  def simpleValidation()(implicit dao: DAO): Boolean = {
-
-    val start = System.currentTimeMillis
-    val validation = CheckpointBlockValidatorNel.validateCheckpointBlock(
-      CheckpointBlock(transactions, checkpoint)
-    )
-    val elapsed = (System.currentTimeMillis - start) / 1000000
-
-    if (validation.isValid) {
-      dao.metrics.incrementMetric("checkpointValidationSuccess")
-    } else {
-      dao.metrics.incrementMetric(Metrics.checkpointValidationFailure)
-      dao.miscLogger.error(s"Checkpoint validation failure: $validation")
-    }
-
-    // TODO: Return Validation instead of Boolean
-    validation.isValid
-  }
-
   def uniqueSignatures: Boolean = signatures.groupBy(_.id).forall(_._2.size == 1)
 
   def signedBy(id: Id): Boolean = witnessIds.contains(id)
@@ -262,323 +241,3 @@ object CheckpointBlock {
   }
 
 }
-
-sealed trait CheckpointBlockValidation {
-
-  def errorMessage: String
-}
-
-case class EmptySignatures() extends CheckpointBlockValidation {
-
-  def errorMessage: String = "CheckpointBlock has no signatures"
-}
-
-case class InvalidSignature(signature: String) extends CheckpointBlockValidation {
-
-  def errorMessage: String = s"CheckpointBlock includes signature=$signature which is invalid"
-}
-
-object InvalidSignature {
-
-  def apply(s: HashSignature) = new InvalidSignature(s.signature)
-}
-
-case class InvalidTransaction(txHash: String) extends CheckpointBlockValidation {
-
-  def errorMessage: String = s"CheckpointBlock includes transaction=$txHash which is invalid"
-}
-
-object InvalidTransaction {
-
-  def apply(t: Transaction) = new InvalidTransaction(t.hash)
-}
-
-case class DuplicatedTransaction(txHash: String) extends CheckpointBlockValidation {
-
-  def errorMessage: String = s"CheckpointBlock includes duplicated transaction=$txHash"
-}
-
-object DuplicatedTransaction {
-
-  def apply(t: Transaction) = new DuplicatedTransaction(t.hash)
-}
-
-case class NoAddressCacheFound(txHash: String, srcAddress: String) extends CheckpointBlockValidation {
-
-  def errorMessage: String =
-    s"CheckpointBlock includes transaction=$txHash which has no address cache for address=$srcAddress"
-}
-
-object NoAddressCacheFound {
-
-  def apply(t: Transaction) = new NoAddressCacheFound(t.hash, t.src.address)
-}
-
-case class InsufficientBalance(address: String, amount: Long, diff: Long) extends CheckpointBlockValidation {
-
-  def errorMessage: String =
-    s"CheckpointBlock includes transaction from address=$address which has insufficient balance"
-}
-
-// TODO: pass also a transaction metadata
-
-case class InternalInconsistency(cbHash: String) extends CheckpointBlockValidation {
-
-  def errorMessage: String =
-    s"CheckpointBlock=$cbHash includes transaction/s which has insufficient balance"
-}
-
-object InternalInconsistency {
-
-  def apply(cb: CheckpointBlock) = new InternalInconsistency(cb.baseHash)
-}
-
-sealed trait CheckpointBlockValidatorNel {
-
-  type ValidationResult[A] = ValidatedNel[CheckpointBlockValidation, A]
-
-  def validateTransactionIntegrity(t: Transaction)(implicit dao: DAO): ValidationResult[Transaction] =
-    if (t.valid) t.validNel else InvalidTransaction(t).invalidNel
-
-  def validateSourceAddressCache(t: Transaction)(implicit dao: DAO): ValidationResult[Transaction] =
-    dao.addressService
-      .lookup(t.src.address)
-      .unsafeRunSync()
-      .fold[ValidationResult[Transaction]](NoAddressCacheFound(t).invalidNel)(_ => t.validNel)
-
-  def validateTransaction(t: Transaction)(implicit dao: DAO): ValidationResult[Transaction] =
-    validateTransactionIntegrity(t)
-//      .product(validateSourceAddressCache(t))
-      .map(_ => t)
-
-  def validateTransactions(
-    t: Iterable[Transaction]
-  )(implicit dao: DAO): ValidationResult[List[Transaction]] =
-    t.toList.map(validateTransaction(_).map(List(_))).combineAll
-
-  def validateDuplicatedTransactions(
-    t: Iterable[Transaction]
-  ): ValidationResult[List[Transaction]] = {
-    val diff = t.toList.diff(t.toSet.toList)
-
-    if (diff.isEmpty) {
-      t.toList.validNel
-    } else {
-
-      def toError(t: Transaction): ValidationResult[Transaction] =
-        DuplicatedTransaction(t).invalidNel
-
-      diff.map(toError(_).map(List(_))).combineAll
-    }
-  }
-
-  def validateSignatureIntegrity(s: HashSignature, baseHash: String): ValidationResult[HashSignature] =
-    if (s.valid(baseHash)) s.validNel else InvalidSignature(s).invalidNel
-
-  def validateSignature(s: HashSignature, baseHash: String): ValidationResult[HashSignature] =
-    validateSignatureIntegrity(s, baseHash)
-      .map(_ => s)
-
-  def validateSignatures(s: Iterable[HashSignature], baseHash: String): ValidationResult[List[HashSignature]] =
-    s.toList.map(validateSignature(_, baseHash).map(List(_))).combineAll
-
-  def validateEmptySignatures(s: Iterable[HashSignature]): ValidationResult[List[HashSignature]] =
-    if (s.nonEmpty) s.toList.validNel else EmptySignatures().invalidNel
-
-  def validateSourceAddressBalances(
-    t: Iterable[Transaction]
-  )(implicit dao: DAO): ValidationResult[List[Transaction]] = {
-
-    def lookup(key: String) =
-      dao.addressService
-        .lookup(key)
-        .unsafeRunSync()
-        .map(_.balance)
-        .getOrElse(0L)
-
-    def validateBalance(address: String, t: Iterable[Transaction]): ValidationResult[List[Transaction]] = {
-      val diff = lookup(address) - t.map(_.amount).sum
-      val amount = t.map(_.amount).sum
-
-      if (diff >= 0L) t.toList.validNel else InsufficientBalance(address, amount, diff).invalidNel
-    }
-
-    t.toList
-      .groupBy(_.src.address)
-      .map(a => validateBalance(a._1, a._2))
-      .toList
-      .combineAll
-  }
-
-  type AddressBalance = Map[String, Long]
-
-  def getParents(c: CheckpointBlock)(implicit dao: DAO): List[CheckpointBlock] = {
-    val parentSOEBaseHashes = c.parentSOEBaseHashes.toList
-
-    if (parentSOEBaseHashes.size != 2) {
-      dao.metrics.incrementMetric("validationParentSOEBaseHashesMissing")
-    }
-
-    val fullData = parentSOEBaseHashes.map(dao.checkpointService.fullData(_).unsafeRunSync())
-
-    if (fullData.exists(_.isEmpty)) {
-      dao.metrics.incrementMetric("validationParentCBLookupMissing")
-    }
-
-    fullData
-      .traverse(_.flatMap(_.checkpointBlock))
-      .getOrElse(List())
-  }
-
-  def isInSnapshot(c: CheckpointBlock)(implicit dao: DAO): Boolean =
-    !dao.snapshotService.acceptedCBSinceSnapshot.get.map(_.contains(c.baseHash)).unsafeRunSync()
-
-  def getSummaryBalance(c: CheckpointBlock)(implicit dao: DAO): AddressBalance = {
-    val spend = c.transactions
-      .groupBy(_.src.address)
-      .mapValues(_.map(-_.amount).sum)
-
-    val received = c.transactions
-      .groupBy(_.dst.address)
-      .mapValues(_.map(_.amount).sum)
-
-    spend |+| received
-  }
-
-  def validateDiff(a: (String, Long))(implicit dao: DAO): Boolean = a match {
-    case (hash, diff) =>
-      dao.addressService
-        .lookup(hash)
-        .unsafeRunSync()
-        .map { _.balanceByLatestSnapshot }
-        .getOrElse(0L) + diff >= 0
-  }
-
-  def validateCheckpointBlockTree(
-    cb: CheckpointBlock
-  )(implicit dao: DAO): Ior[NonEmptyList[CheckpointBlockValidation], AddressBalance] =
-    if (isInSnapshot(cb)) Map.empty[String, Long].rightIor
-    else
-      getParents(cb)
-        .map(validateCheckpointBlockTree)
-        .foldLeft(Map.empty[String, Long].rightIor[NonEmptyList[CheckpointBlockValidation]])(
-          (result, d) => result.combine(d)
-        )
-        .map(getSummaryBalance(cb) |+| _)
-        .flatMap(
-          diffs =>
-            if (diffs.forall(validateDiff))
-              diffs.rightIor
-            else
-              Ior.both(NonEmptyList.of(InternalInconsistency(cb)), diffs)
-        )
-
-  implicit def validateTreeToValidated(
-    v: Ior[NonEmptyList[CheckpointBlockValidation], AddressBalance]
-  ): ValidationResult[AddressBalance] =
-    v match {
-      case Ior.Right(a)   => a.validNel
-      case Ior.Left(a)    => a.invalid
-      case Ior.Both(a, _) => a.invalid
-    }
-
-  def validateCheckpointBlock(
-    cb: CheckpointBlock
-  )(implicit dao: DAO): ValidationResult[CheckpointBlock] = {
-    val preTreeResult =
-      validateEmptySignatures(cb.signatures)
-        .product(validateSignatures(cb.signatures, cb.baseHash))
-        .product(validateTransactions(cb.transactions))
-        .product(validateDuplicatedTransactions(cb.transactions))
-        .product(validateSourceAddressBalances(cb.transactions))
-
-    val postTreeIgnoreEmptySnapshot =
-      if (dao.snapshotService.lastSnapshotHeight.get.unsafeRunSync() == 0) preTreeResult
-//      else preTreeResult.product(validateCheckpointBlockTree(cb))
-      else preTreeResult
-
-    postTreeIgnoreEmptySnapshot.map(_ => cb)
-  }
-
-  def getTransactionsTillSnapshot(
-    cbs: Seq[CheckpointBlock]
-  )(implicit dao: DAO): Seq[String] = {
-
-    @tailrec
-    def getParentTransactions(
-      parents: Seq[CheckpointBlock],
-      accu: Seq[String] = Seq.empty,
-      snapshotReached: Boolean = false
-    ): Seq[String] =
-      parents match {
-        case Nil => accu
-        case cb :: tail if !snapshotReached && !isInSnapshot(cb) =>
-          getParentTransactions(
-            tail ++ getParents(cb),
-            accu ++ cb.transactions.map(_.hash),
-            isInSnapshot(cb)
-          )
-        case cb :: tail if snapshotReached || isInSnapshot(cb) =>
-          getParentTransactions(
-            tail,
-            accu ++ cb.transactions.map(_.hash),
-            snapshotReached
-          )
-      }
-
-    cbs
-      .filterNot(isInSnapshot)
-      .flatMap(cb => getParentTransactions(getParents(cb), cb.transactions.map(_.hash)))
-  }
-
-  def containsAlreadyAcceptedTx(cb: CheckpointBlock)(implicit dao: DAO): IO[List[String]] = {
-    val containsAccepted = cb.transactions.toList.map { t =>
-      dao.transactionService.lookup(t.hash).map {
-        case Some(tx) if tx.cbBaseHash != cb.baseHash.some => (t.hash, true)
-        case _                                             => (t.hash, false)
-      }
-      dao.transactionService.isAccepted(t.hash).map(b => (t.hash, b))
-    }.sequence[IO, (String, Boolean)]
-      .map(l => l.collect { case (h, true) => h })
-
-    containsAccepted
-  }
-
-  def detectInternalTipsConflict(
-    inputTips: Seq[CheckpointCache]
-  )(implicit dao: DAO): Option[CheckpointCache] = {
-
-    def detect(
-      tips: Seq[CheckpointCache],
-      ancestryTransactions: Map[Transaction, String] = Map.empty
-    ): Option[CheckpointCache] =
-      tips match {
-        case Nil => None
-        case CheckpointCache(Some(cb), children, height) :: _
-            if cb.transactions.toSet.intersect(ancestryTransactions.keySet).nonEmpty =>
-          val conflictingCBBaseHash = ancestryTransactions(
-            cb.transactions.toSet.intersect(ancestryTransactions.keySet).head
-          )
-          Some(
-            selectBlockToPreserve(
-              Seq(CheckpointCache(Some(cb), children, height)) ++ inputTips
-                .filter(ccd => ccd.checkpointBlock.exists(_.baseHash == conflictingCBBaseHash))
-            )
-          )
-        case CheckpointCache(Some(cb), _, _) :: tail =>
-          detect(tail, ancestryTransactions ++ cb.transactions.map(i => (i, cb.baseHash)))
-      }
-    detect(inputTips)
-  }
-
-  def hasTransactionInCommon(cbLeft: CheckpointBlock, cbRight: CheckpointBlock): Boolean =
-    cbLeft.transactions.intersect(cbRight.transactions).nonEmpty
-
-  def selectBlockToPreserve(blocks: Iterable[CheckpointCache]): CheckpointCache =
-    blocks.maxBy(
-      cb => (cb.children, cb.checkpointBlock.map(b => (b.signatures.size, b.baseHash)))
-    )
-
-}
-
-object CheckpointBlockValidatorNel extends CheckpointBlockValidatorNel
