+++ src/test/scala/org/constellation/TestHelpers.scala
@@ -6,6 +6,7 @@ import better.files.File
 import cats.effect.IO
 import com.google.common.hash.Hashing
 import com.typesafe.scalalogging.Logger
+import org.constellation.checkpoint.CheckpointService
 import org.constellation.consensus.ConsensusRemoteSender
 import org.constellation.crypto.KeyUtils
 import org.constellation.crypto.KeyUtils.makeKeyPair
