+++ src/test/scala/org/constellation/checkpoint/CheckpointBlockValidatorNelTest.scala
@@ -1,22 +1,22 @@
-package org.constellation.primitives
+package org.constellation.checkpoint
 
 import java.security.KeyPair
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import akka.testkit.{TestKit, TestProbe}
+import akka.testkit.TestKit
 import cats.effect.{ContextShift, IO}
-import cats.effect.concurrent.Ref
 import cats.implicits._
 import constellation.createTransaction
 import io.chrisdavenport.log4cats.SelfAwareStructuredLogger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
+import org.constellation.checkpoint.CheckpointBlockValidator._
 import org.constellation.consensus.{RandomData, Snapshot, SnapshotInfo}
 import org.constellation.p2p.Cluster
-import org.constellation.primitives.CheckpointBlockValidatorNel._
-import org.constellation.primitives.Schema.{AddressCacheData, CheckpointCache, Height, Id}
+import org.constellation.primitives.Schema.{AddressCacheData, CheckpointCache, Id}
 import org.constellation.primitives.concurrency.SingleRef
-import org.constellation.storage.{CheckpointBlocksMemPool, CheckpointService, SnapshotService, TransactionService}
+import org.constellation.primitives.{CheckpointBlock, IPManager, Transaction}
+import org.constellation.storage._
 import org.constellation.util.{HashSignature, Metrics}
 import org.constellation.{ConstellationExecutionContext, DAO, NodeConfig}
 import org.mockito.cats.IdiomaticMockitoCats
@@ -37,6 +37,10 @@ class CheckpointBlockValidatorNelTest
 
   val snapService: SnapshotService[IO] = mock[SnapshotService[IO]]
   val checkpointService: CheckpointService[IO] = mock[CheckpointService[IO]]
+  val addressService: AddressService[IO] = mock[AddressService[IO]]
+
+  val checkpointBlockValidator: CheckpointBlockValidator[IO] =
+    new CheckpointBlockValidator[IO](addressService, snapService, checkpointService, dao)
 
   val leftBlock: CheckpointBlock = mock[CheckpointBlock]
   val leftParent: CheckpointBlock = mock[CheckpointBlock]
@@ -105,7 +109,7 @@ class CheckpointBlockValidatorNelTest
   }
 
   test("it should detect no conflict and return None") {
-    containsAlreadyAcceptedTx(leftBlock).unsafeRunSync() shouldBe List.empty
+    checkpointBlockValidator.containsAlreadyAcceptedTx(leftBlock).unsafeRunSync() shouldBe List.empty
   }
 
   test("it should detect direct internal conflict with other tip") {
@@ -129,7 +133,7 @@ class CheckpointBlockValidatorNelTest
     val conflictTx = leftBlock.transactions.head.hash
     dao.transactionService.isAccepted(conflictTx) shouldReturn IO.pure(true)
 
-    containsAlreadyAcceptedTx(leftBlock).unsafeRunSync() shouldBe List(conflictTx)
+    checkpointBlockValidator.containsAlreadyAcceptedTx(leftBlock).unsafeRunSync() shouldBe List(conflictTx)
   }
 
   test("it should get transactions from parent") {
@@ -139,13 +143,13 @@ class CheckpointBlockValidatorNelTest
     checkpointService.fullData("ancestor_in_snap") shouldReturn IO.pure(None)
 
     val combinedTxs =
-      getTransactionsTillSnapshot(List(rightBlock))
+      checkpointBlockValidator.getTransactionsTillSnapshot(List(rightBlock)).unsafeRunSync()
     combinedTxs shouldBe (rightBlock.transactions ++ rightParent.transactions).map(_.hash)
   }
 
   test("it should return false for cb not in snap") {
-    isInSnapshot(rightParent) shouldBe false
-    isInSnapshot(leftParent) shouldBe false
+    checkpointBlockValidator.isInSnapshot(rightParent).unsafeRunSync() shouldBe false
+    checkpointBlockValidator.isInSnapshot(leftParent).unsafeRunSync() shouldBe false
   }
 
   test("it should return correct block to preserve with greater base hash") {
@@ -191,6 +195,9 @@ class ValidationSpec
   implicit val logger: SelfAwareStructuredLogger[IO] = Slf4jLogger.getLogger[IO]
   implicit val timer = IO.timer(ConstellationExecutionContext.unbounded)
   implicit val cs = IO.contextShift(ConstellationExecutionContext.bounded)
+
+  val checkpointBlockValidator: CheckpointBlockValidator[IO] =
+    new CheckpointBlockValidator[IO](dao.addressService, dao.snapshotService, dao.checkpointService, dao)
 
   val ipManager = IPManager[IO]()
   val cluster = Cluster[IO](() => dao.metrics, ipManager, dao)
@@ -281,7 +288,7 @@ class ValidationSpec
           .foreach(cb => cb.store(CheckpointCache(Some(cb))))
 
         println(dao.metrics)
-        assert(cb7.simpleValidation())
+        assert(checkpointBlockValidator.simpleValidation(cb7).unsafeRunSync().isValid)
       }
     }
 
@@ -304,7 +311,7 @@ class ValidationSpec
           )
         )
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
 
@@ -324,7 +331,7 @@ class ValidationSpec
           )
         )
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
 
@@ -347,7 +354,7 @@ class ValidationSpec
           )
         )
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
 
@@ -370,7 +377,7 @@ class ValidationSpec
           )
         )
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
 
@@ -386,7 +393,7 @@ class ValidationSpec
 
         val cb = CheckpointBlock.createCheckpointBlockSOE(txs, startingTips)
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
 
@@ -408,7 +415,7 @@ class ValidationSpec
           )
         )
 
-        assert(!cb.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb).unsafeRunSync().isValid)
       }
     }
   }
@@ -433,14 +440,14 @@ class ValidationSpec
         val tx2 = createTransaction(getAddress(a), getAddress(c), 75L, a)
         val cb2 = CheckpointBlock.createCheckpointBlockSOE(Seq(tx2), startingTips)
 
-        if (cb1.simpleValidation()) {
+        if (checkpointBlockValidator.simpleValidation(cb1).unsafeRunSync().isValid) {
           cb1.transactions.toList
             .map(dao.addressService.transfer)
             .sequence[IO, AddressCacheData]
             .unsafeRunSync()
         }
 
-        assert(!cb2.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb2).unsafeRunSync().isValid)
       }
     }
   }
@@ -508,10 +515,10 @@ class ValidationSpec
         Seq(cb1, cb2, cb3, cb4, cb5, cb6, cb7).foreach { cb =>
           cb.store(CheckpointCache(cb.some))
 //          // TODO: wkoszycki #420 enable recursive validation transactions shouldn't be required to be stored to run validation
-          dao.checkpointService.acceptTransactions(cb).unsafeRunSync()
+          dao.checkpointAcceptanceService.acceptTransactions(cb).unsafeRunSync()
         }
 
-        assert(!cb7.simpleValidation())
+        assert(!checkpointBlockValidator.simpleValidation(cb7).unsafeRunSync().isValid)
       }
     }
   }
