+++ src/test/scala/org/constellation/primitives/TipServiceTest.scala
@@ -3,9 +3,10 @@ import cats.effect.{ContextShift, IO, Timer}
 import cats.implicits._
 import com.typesafe.scalalogging.Logger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
+import org.constellation.checkpoint.CheckpointService
 import org.constellation.consensus.RandomData
 import org.constellation.primitives.Schema.{CheckpointCacheMetadata, EdgeHashType, Height, TypedEdgeHash}
-import org.constellation.storage.{CheckpointService, SOEService}
+import org.constellation.storage.SOEService
 import org.constellation.util.Metrics
 import org.constellation.{ConstellationExecutionContext, DAO, Fixtures}
 import org.mockito.{ArgumentMatchersSugar, IdiomaticMockito}
