+++ src/test/scala/org/constellation/storage/SnapshotServiceTest.scala
@@ -4,6 +4,7 @@ import better.files.File
 import cats.effect.{ContextShift, IO, Timer}
 import org.constellation._
 import cats.implicits._
+import org.constellation.checkpoint.CheckpointService
 import org.constellation.consensus.{ConsensusManager, RandomData, Snapshot, SnapshotInfo}
 import org.constellation.primitives.Schema.{CheckpointCache, Id, NodeState}
 import org.constellation.primitives.ConcurrentTipService
